Cleanup in ffmpeg, fileformar filecr2, file{exr,jpeg,png,tga}
Removed out commented printfs, unused parts of source
Implemented framenum, samplenum. A lot of cosmetics.
Removed unused filejpeglist
Compilation fix in maincursor

+++ cinelerra/maincursor.C
@@ -76,7 +76,7 @@ void MainCursor::deactivate()
 	}
 }
 
-int MainCursor::repeat_event(int64_t duration)
+int MainCursor::repeat_event(int duration)
 {
 	if(!active || !gui->get_has_focus()) return 0;
 	if(duration != BC_WindowBase::get_resources()->blink_rate) return 0;
